// React/app/GerAdv_TS/ProcessOutPutIDs/Interfaces/interface.ProcessOutPutIDs.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IProcessOutPutIDs {
// 202501251
    id: number;
 
	nome: string,
}

export interface IProcessOutPutIDsFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IProcessOutPutIDsIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}